requirements.txt
gunicorn
django-heroku
whitenoise
python-dotenv
django-cors-headers
pillow
cloudinary